css files refactored
