+++ frontend/src/assets/css/food-page.css
@@ -1,32 +1,3 @@
-.app-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    max-height: 80vh;
-    padding-top: 120px;
-}
-
-.food-header {
-    position: relative;
-    width: 100%;
-    display: flex;
-    align-items: flex-start;
-    justify-content: center;
-    gap: 20px;
-}
-
-.navigation-buttons {
-    position: fixed;
-    top: 20px;
-    left: 20px;
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    z-index: 1000;
-}
-
 .content-container.two-column {
     display: flex;
     flex-direction: row;
@@ -101,54 +72,12 @@
     transform: rotate(-45deg);
 }
 
-.food-ingredients-wrapper {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-top: 60px;
-}
-
 .ingredients-table-wrapper {
     display: flex;
     justify-content: center;
     align-items: center;
     width: 100%;
     padding: 20px;
-}
-
-.ingredient-icons {
-    display: flex;
-    gap: 10px;
-    align-items: center;
-    justify-content: flex-end;
-}
-
-.ingredient-row {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 100%;
-}
-
-.ingredient-label-wrapper {
-    display: flex;
-    align-items: baseline;
-    gap: 8px;
-    flex: 1;
-    white-space: nowrap; /* Prevent wrapping */
-}
-
-.ingredient-name {
-    font-size: 1.1rem;
-    display: inline-block;
-    white-space: nowrap;
-}
-
-.ingredient-quantity {
-    font-size: 0.9rem;
-    opacity: 0.8;
-    margin-left: 2px;
-    white-space: nowrap;
 }
 
 .editable-description::-webkit-scrollbar {
