+++ frontend/src/assets/css/meal-plan-vertical.css
@@ -1,14 +1,3 @@
-.app-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    max-height: 80vh;
-    padding-top: 120px;
-}
-
-
 @media (max-width: 768px) {
     .meal-view-container {
         width: 95%;
@@ -25,7 +14,7 @@
     }
 }
 
-.mealtime-table-wrapper {
+.vertical-table-wrapper {
     display: flex;
     justify-content: center;
     align-items: center;
@@ -33,7 +22,7 @@
     width: 100%;
 }
 
-.mealtime-container {
+.vertical-table-container {
     text-align: center;
     width: 100%;
     display: flex;
@@ -41,7 +30,7 @@
     align-items: center;
 }
 
-.mealtime-header {
+.meal-plan-item-header {
     position: fixed;
     top: 20px;
     left: 0;
@@ -52,14 +41,4 @@
     padding: 0 20px;
     z-index: 11;
     gap: 20px;
-}
-
-.mealtime-buttons {
-    position: fixed;
-    top: 20px;
-    left: 20px;
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    z-index: 1000;
 }
