--- frontend/src/assets/css/pages/day-page.css
@@ -1,65 +1,0 @@
-.app-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    max-height: 80vh;
-    padding-top: 120px;
-}
-
-
-@media (max-width: 768px) {
-    .meal-view-container {
-        width: 95%;
-    }
-
-    .meal-view-table {
-        border-spacing: 5px;
-        font-size: 14px;
-    }
-
-    .modal-content {
-        width: 90%;
-        margin: 10% auto;
-    }
-}
-
-.day-table-wrapper {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-grow: 1;
-    width: 100%;
-}
-
-.day-container {
-    text-align: center;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.day-header {
-    position: fixed;
-    top: 20px;
-    left: 0;
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: flex-start;
-    padding: 0 20px;
-    z-index: 11;
-    gap: 20px;
-}
-
-.day-buttons {
-    position: fixed;
-    top: 20px;
-    left: 20px;
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    z-index: 1000;
-}
