--- frontend/src/assets/css/pages/foods-page.css
@@ -1,81 +1,0 @@
-.app-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    max-height: 80vh;
-    padding-top: 120px;
-}
-
-.foods-header {
-    position: relative;
-    width: 100%;
-    display: flex;
-    align-items: flex-start;
-    justify-content: center;
-    gap: 20px;
-}
-
-.menu-button {
-    position: fixed;
-    top: 20px;
-    left: 20px;
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    z-index: 1000;
-}
-
-.edit-buttons {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: flex-end;
-    gap: 10px;
-}
-
-.edit-button {
-    cursor: pointer;
-    width: 20px;
-    height: 20px;
-    display: block;
-}
-
-.foods-wrapper {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-top: 60px;
-}
-
-.foods-table-wrapper {
-    display: flex;
-    justify-content: center;
-    text-align: center;
-    align-items: center;
-    flex-direction: column;
-}
-
-.food-row {
-    position: relative;
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-}
-
-.food-name {
-    position: absolute;
-    left: 50%;
-    transform: translate(-50%);
-    font-size: 1.1rem;
-    cursor: pointer;
-}
-
-.food-icons {
-    display: flex;
-    gap: 10px;
-    align-items: center;
-    justify-content: flex-end;
-}
