--- frontend/src/assets/css/pages/ingredients-at-home-page.css
@@ -1,89 +1,0 @@
-.app-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    max-height: 80vh;
-    padding-top: 120px;
-}
-
-.home-ingredients-header {
-    position: relative;
-    width: 100%;
-    display: flex;
-    align-items: flex-start;
-    justify-content: center;
-    gap: 20px;
-}
-
-.menu-button {
-    position: fixed;
-    top: 20px;
-    left: 20px;
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    z-index: 1000;
-}
-
-.edit-buttons {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: flex-end;
-    gap: 10px;
-}
-
-.home-ingredients-wrapper {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-top: 60px;
-}
-
-.home-ingredients-table-wrapper {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-grow: 1;
-    width: 100%;
-    padding: 20px;
-    max-width: 95vw;
-}
-
-.home-ingredient-icons {
-    display: flex;
-    gap: 10px;
-    align-items: center;
-    justify-content: flex-end;
-}
-
-.home-ingredient-row {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 100%;
-}
-
-.home-ingredient-label-wrapper {
-    display: flex;
-    align-items: baseline;
-    gap: 8px;
-    flex: 1;
-    white-space: nowrap; /* Prevent wrapping */
-}
-
-.home-ingredient-name {
-    font-size: 1.1rem;
-    display: inline-block;
-    white-space: nowrap;
-}
-
-.home-ingredient-quantity {
-    font-size: 0.9rem;
-    opacity: 0.8;
-    margin-left: 2px;
-    white-space: nowrap;
-}
-
