--- frontend/src/assets/css/pages/shopping-list-page.css
@@ -1,94 +1,0 @@
-.app-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    max-height: 80vh;
-    padding-top: 120px;
-}
-
-.shopping-list-header {
-    position: relative;
-    width: 100%;
-    display: flex;
-    align-items: flex-start;
-    justify-content: center;
-    gap: 20px;
-}
-
-.menu-buttons {
-    position: fixed;
-    top: 20px;
-    left: 20px;
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    z-index: 1000;
-}
-
-.shopping-list-wrapper {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-top: 60px;
-}
-
-.shopping-list-table-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-grow: 1;
-    width: 100%;
-    padding: 20px;
-    max-width: 95vw;
-}
-
-.shopping-item-row {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 100%;
-}
-
-.shopping-item-amount {
-    font-size: 0.9rem;
-    opacity: 0.8;
-    margin-left: 2px;
-    white-space: nowrap;
-}
-
-.shopping-item-price {
-    font-size: 0.9rem;
-    opacity: 0.8;
-    margin-left: 2px;
-    white-space: nowrap;
-}
-
-.shopping-item-name {
-    font-size: 1.1rem;
-    display: inline-block;
-    white-space: nowrap;
-}
-
-.shopping-item-icons {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: flex-end;
-    gap: 10px;
-}
-
-.edit-button {
-    cursor: pointer;
-    width: 20px;
-    height: 20px;
-    display: block;
-}
-
-.shopping-item-label {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: 10px;
-}
