+++ frontend/src/assets/css/shopping-list-page.css
@@ -1,0 +1,36 @@
+.shopping-list-wrapper {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-top: 60px;
+}
+
+.shopping-list-table-container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-grow: 1;
+    width: 100%;
+    padding: 20px;
+    max-width: 95vw;
+}
+
+.shopping-item-name {
+    font-size: 1.1rem;
+    display: inline-block;
+    white-space: nowrap;
+}
+
+.shopping-item-param {
+    font-size: 0.9rem;
+    opacity: 0.8;
+    margin-left: 2px;
+    white-space: nowrap;
+}
+
+.shopping-item-label {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 10px;
+}
