+++ frontend/src/pages/FoodsPage.jsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { useNavigate} from "react-router-dom";
 import { useFoods} from "../hooks/useFoods";
-import '../assets/css/pages/foods-page.css';
 import PageTitle from "../components/PageTitle";
 import TapeButton from "../components/TapeButton";
 import TapedTable from "../components/TapedTable";
@@ -33,55 +32,53 @@ const FoodsPage = () => {
 
     return (
         <div className="app-container">
-            <div className="foods-header">
-                <div className="menu-button">
+            <div className="page-header">
+                <div className="header-buttons">
                     <TapeButton text="Menu" onClick={navigateToMenu} />
                 </div>
                 <PageTitle text="Foods"/>
             </div>
-            <div className="foods-wrapper">
-                <div className="foods-table-container">
-                    <TapedTable
-                        layout="vertical"
-                        rows={sortedFoods}
-                        renderCell={(rowIndex) => {
-                            const f = sortedFoods[rowIndex];
-                            return (
-                                <div className="food-row">
-                                    <span className="food-name" onClick={() => navigate(`/food/${encodeURIComponent(f.name)}`)}>{f.name}</span>
-                                    <div className="food-icons">
-                                        <div className="edit-buttons">
-                                            <img
-                                                src={require('../assets/images/pencil.png')}
-                                                alt="Edit"
-                                                className="edit-button"
-                                                onClick={() => handleEditFood(f)}
-                                            />
-                                            <img
-                                                src={require('../assets/images/trashcan.png')}
-                                                alt="Delete"
-                                                className="edit-button"
-                                                onClick={() => handleDeleteFood(f.id)}
-                                            />
-                                        </div>
+            <div className="content-wrapper">
+                <TapedTable
+                    layout="vertical"
+                    rows={sortedFoods}
+                    renderCell={(rowIndex) => {
+                        const f = sortedFoods[rowIndex];
+                        return (
+                            <div className="table-row">
+                                <span className="cell-name" onClick={() => navigate(`/food/${encodeURIComponent(f.name)}`)}>{f.name}</span>
+                                <div className="cell-icons">
+                                    <div className="edit-buttons">
+                                        <img
+                                            src={require('../assets/images/pencil.png')}
+                                            alt="Edit"
+                                            className="edit-button"
+                                            onClick={() => handleEditFood(f)}
+                                        />
+                                        <img
+                                            src={require('../assets/images/trashcan.png')}
+                                            alt="Delete"
+                                            className="edit-button"
+                                            onClick={() => handleDeleteFood(f.id)}
+                                        />
                                     </div>
                                 </div>
-                            );
-                        }}
-                        allowReorder={false}
-                        showHeader={false}
-                        showRowLabels={false}
-                        extraBottomRow={
-                            <tr>
-                                <td>
-                                    <button className="table-button lobster" onClick={() => setIsModalOpen(true)}>
-                                        Add new food
-                                    </button>
-                                </td>
-                            </tr>
-                        }
-                    />
-                </div>
+                            </div>
+                        );
+                    }}
+                    allowReorder={false}
+                    showHeader={false}
+                    showRowLabels={false}
+                    extraBottomRow={
+                        <tr>
+                            <td>
+                                <button className="table-button lobster" onClick={() => setIsModalOpen(true)}>
+                                    Add new food
+                                </button>
+                            </td>
+                        </tr>
+                    }
+                />
                 <NewNameModal
                     isOpen={isModalOpen}
                     onClose={() => setIsModalOpen(false)}
